+++ PressureRack/DruckState.cs
@@ -11,17 +11,17 @@ namespace PressureRack
     {
         public bool State { get; private set; }
         public string SN { get; private set; }
-        public float RangeLo { get; private set; }
-        public float RangeHi { get; private set; }
-        public float PrecClass { get; private set; }
+        public double RangeLo { get; private set; }
+        public double RangeHi { get; private set; }
+        public double PrecClass { get; private set; }
         
 
         public DruckState(string str) // Декодирует принятую из сокета строку в информацию
         {
             State = (Decoder.ExtractData("STATE:", str) == "ON");
-            RangeLo = Convert.ToSingle(Decoder.ExtractData("LO:", str));
-            RangeHi = Convert.ToSingle(Decoder.ExtractData("HI:", str));
-            PrecClass = Convert.ToSingle(Decoder.ExtractData("PREC:", str));
+            RangeLo = Convert.ToDouble(Decoder.ExtractData("LO:", str));
+            RangeHi = Convert.ToDouble(Decoder.ExtractData("HI:", str));
+            PrecClass = Convert.ToDouble(Decoder.ExtractData("PREC:", str));
             SN = Decoder.ExtractData("SN:", str);
         }
     }
